+++ semcore/date-picker/src/MonthPicker.jsx
@@ -52,8 +52,16 @@ class MonthPickerRoot extends PickerAbstract {
   }
 
   getTriggerProps() {
-    const { value, onChange, onDisplayedPeriodChange, locale, disabled, size, getI18nText } =
-      this.asProps;
+    const {
+      value,
+      onChange,
+      onDisplayedPeriodChange,
+      locale,
+      disabled,
+      disabledErrorText,
+      size,
+      getI18nText,
+    } = this.asProps;
 
     return {
       ...super.getButtonTriggerProps(),
@@ -64,6 +72,7 @@ class MonthPickerRoot extends PickerAbstract {
       w: size === 'm' ? 120 : 135,
       parts: dateParts,
       disabledDates: disabled,
+      disabledErrorText,
       children: () => <InputTrigger.SingleDateInput />,
       getI18nText,
     };