+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.26.2] - 2024-04-11
+
+### Fixed
+
+- DropdownMenu with input in the trigger was not highting current items when focus comes to input with mouse click.
+
+## [4.26.1] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/dropdown` [4.23.1 ~> 4.23.2], `@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [4.26.0] - 2024-04-03
+
+### Added
+
+- `DropdownMenu.Nesting.Item` for more granular control over nested dropdown menus.
+
+## [4.25.1] - 2024-04-03
+
+### Fixed
+
+- Menu item was displaying highlighted ring when dropdown is opened by mouse interaction.
+
 ## [4.25.0] - 2024-03-27
 
 ### Changed