+++ semcore/textarea/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.21.2] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/flex-box` [5.21.1 ~> 5.21.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [5.21.1] - 2024-04-02
+
+### Fixed
+
+- Copying a textarea node with computed styles for calculating rows.
+
 ## [5.21.0] - 2024-03-27
 
 ### Changed